Move candidate parsing logic to the Candidate class
Implements a new static method `Candidate::ParseCandidateString` on the
`Candidate` class. This method handles parsing an RFC 5245 ICE candidate
string and returns an `RTCErrorOr<Candidate>` to provide error
information. ParseCandidateString() is the counterpart to the existing
ToCandidateAttribute() method.

To ensure a safe migration, the existing parser in `webrtc_sdp.cc` is
temporarily modified. to ensure that ParseCandidateString() functions identically to the legacy ParseCandidate() method, which will be replaced in a follow-up CL.

Bug: webrtc:12330
Change-Id: Ie979b3c944309e6a49462fb65b04b428062e38fd
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/411182
Reviewed-by: Harald Alvestrand <hta@webrtc.org>
Commit-Queue: Tomas Gunnarsson <tommi@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#45799}