[owners] Remove tkchin@webrtc.org from api/OWNERS
This suggested change is automatically generated based on commit
and review activity to improve the reliability of OWNERS suggestions
and CL review latency.

If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.
If no CR -1 vote is set after 5 days, the CL will be submitted.
OWNERS expectations for the chromium project are at:
https://chromium.googlesource.com/chromium/src/+/refs/heads/main/docs/code_reviews.md#expectations-of-owners

See the owner's recent activity for context:
https://webrtc-review.googlesource.com/q/tkchin@webrtc.org

To report an issue, file a bug in the Infra>Codereview component.

No-try: true
Bug: none
Change-Id: I7fdf6701d2b41ed95659a3039b9445d85d6e1a1f
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/415320
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Tomas Gunnarsson <tommi@webrtc.org>
Commit-Queue: Tomas Gunnarsson <tommi@webrtc.org>
Reviewed-by: Harald Alvestrand <hta@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#45875}