sdp: detect modification of b= line
which does not even have an effect since b= in the local
description is asking the remote side to please not send
more than the amount specified. It does not control what
the local side uses as bandwidth limit which can be
configured with setParameters.

Bug: webrtc:414284082
Change-Id: Iefef7a7b5b48d329509063b16eb4f2679078fbd3
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/415740
Reviewed-by: Harald Alvestrand <hta@webrtc.org>
Commit-Queue: Philipp Hancke <philipp.hancke@googlemail.com>
Reviewed-by: Henrik Boström <hbos@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#45897}