PixelLimitResource: Add parameter for adaptation reason.
The "reason" for adapting, which can be "cpu" or "quality" depending on
what the resource is reportedly monitoring, impacts which
RTCQualityAdaptationReason[1] we report in getStats() where these are
mapped to "cpu" and "bandwidth".

For testing purposes, it's useful to be able to control if
PixelLimitResource results in being CPU or BW limited. This CL adds the
new field trial enum parameter, "reason".

In addition to unit tests in this CL I have manually verified this gets
correctly plumbed to getStats() by going to
https://henbos.github.io/codec-quality/src/index.html using
--force-fieldtrials=WebRTC-PixelLimitResource/target_pixels:64800,interval:1s,reason:quality,toggle:10s/

[1] https://w3c.github.io/webrtc-stats/#rtcqualitylimitationreason-enum

Bug: webrtc:447037084
Change-Id: I575e1a984a6e31a997ba43f3565b709465fafea5
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/413160
Commit-Queue: Henrik Boström <hbos@webrtc.org>
Reviewed-by: Ilya Nikolaevskiy <ilnik@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#45896}