Move webrtc_sdp serialization methods and SessionDescriptionInterface
...to the jsep target.

This resolves the circular dependency by moving the definitions and implementation of the SDP serialization and deserialization functions and `SessionDescriptionInterface` functions from `pc/` to `api/` and including them with the `jsep` target and in alignment with the JSEP API header.

The webrtc_sdp build target lives on for a bit longer while there may be downstream code that includes the webrtc_sdp.h header.

Bug: webrtc:42222470
Change-Id: If1417455729e128361575f2da161444bfed8b21e
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/415745
Reviewed-by: Harald Alvestrand <hta@webrtc.org>
Commit-Queue: Tomas Gunnarsson <tommi@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#45909}