Implement `DecodabilityTracker`.
The `DecodabilityTracker` is responsible for taking a sequence of assembled frames and producing a sequence of decodable frames.

The implementation currently skips over all reordered frames. This will be improved upon later.

In upcoming CLs, the `DecodabilityTracker` will be integrated into an RtcEventLog-driven simulator.

Bug: b/423646186
Change-Id: I1d4edeb7b2efe8a76183fdce08044eafe96d0255
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/415520
Reviewed-by: Åsa Persson <asapersson@webrtc.org>
Commit-Queue: Rasmus Brandt <brandtr@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#45917}