Make `Assembler` non-API interface implementations private
The `Assembler` implements the `Transport` and
`RtpVideoStreamReceiver2::OnCompleteFrameCallback` interfaces. Those
interfaces are called on by the dependencies of the class, and thus do
not belong to the public API. This is made clear in this CL by moving
the declarations to private.

We could have also hid these implementations in a sub-class, but that is
unnecessarily complex for this simple use case.

Bug: b/423646186
Change-Id: Iab5967c5fe42d33be313c3c42d7257182067367c
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/416620
Auto-Submit: Rasmus Brandt <brandtr@webrtc.org>
Reviewed-by: Åsa Persson <asapersson@webrtc.org>
Commit-Queue: Rasmus Brandt <brandtr@webrtc.org>
Commit-Queue: Åsa Persson <asapersson@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#45918}