Remove the VideoReceiveChannel::SetReceiverFeedbackParameters function
this was called from a callback whenever the send codec changed, but
investigation showed that it never changed anything as long as it
was only called after configuring the receive channel; all config
was taken from the SDP in any case.

Tested by adding CHECKs into the function that verified that no change
was ever attempted except for tests that targeted this function only.
Those tests are also deleted.

Bug: webrtc:42224170
Change-Id: I5e1e9dca06be4606b94d960d6ce006c06066c28c
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/416940
Reviewed-by: Henrik Boström <hbos@webrtc.org>
Commit-Queue: Harald Alvestrand <hta@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#45943}