Implement `RenderingTracker`.
The `RenderingTracker` is responsible for taking a sequence of assembled
frames and producing a sequence of rendered frames.

The implementation relies on the production classes
`VideoStreamBufferController` and `IncomingVideoStream` for the jitter
buffering and render buffering, respectively.

In upcoming CLs, the `RenderingTracker` will be integrated into an
RtcEventLog-driven simulator.

Bug: b/423646186
Change-Id: I5ab7bb6c8b4b21c024f439e9956c8e4068e9c782
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/416720
Reviewed-by: Åsa Persson <asapersson@webrtc.org>
Commit-Queue: Rasmus Brandt <brandtr@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#45962}