Add ability to re-enable header extensions in 2nd O/A
This ensures that disabling an RTP header extension in a first
O/A and then enabling it in a second O/A works as expected.

It preserves the ability to reject an RTP header extension
between an incoming offer and an outgoing answer by using
SetHeaderExtensionsToNegotiate.

Bug: webrtc:439514253
Change-Id: Ie8a8946f8c7970b91858b16b48b62702c7dea0ea
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/417963
Commit-Queue: Harald Alvestrand <hta@webrtc.org>
Reviewed-by: Per Kjellander <perkj@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#45968}