Implement `RtcEventLogDriver`.
The `RtcEventLogDriver` is the simulation orchestrator. It takes a
`ParsedRtcEventLog`, sets up the simulation environment, and iterates
over all relevant events in the log. As the log is iterated, the
simulation clock is advanced accordingly.

The `RtcEventLogDriver` will be used by the upcoming
`{Decodability,Rendering}Simulator`s. Those classes will be responsible
for abstracting away the specifics of the simulation, by implementing
the `RtcEventLogDriver::StreamInterface`.

Bug: b/423646186
Change-Id: I9bd104a1f6234bd50c809f3b97c9832ed4e62994
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/416721
Reviewed-by: Åsa Persson <asapersson@webrtc.org>
Commit-Queue: Rasmus Brandt <brandtr@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#45997}