DTLS1.3: Fix missing ACK sending due to failure to set timer.
DTLS1.3 sometimes "restart" the timer, i.e.
it was first disabled and then is later turned on.
For this case, the failure to send the ACK resulted
in other end to exhaust the retries ultimately
to disconnect the DTLS connection.

This patch modifies so that we check/set the timer
after each SSL_{read/write/connect} call.

This fix contains a testcase that fails before this patch
and works afterwards. The testcase only failed for dtls1.3 and
only when running for longer than 6 minutes, which is when the
retransmit counter is exhausted.

In "real life" logs I estimated the problem happen around 6 minutes :)
This problem was found in the wild with real application that
started experiencing too high percentage of connections
being broken. The problem occurs if last packet in handshake is lost,
but can probably happens in other scenarios too.

Bug: chromium:441245658
Change-Id: Ie461db8889c5a196eadca8287bd6334dd207016a
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/419380
Reviewed-by: Harald Alvestrand <hta@webrtc.org>
Commit-Queue: Jonas Oreland <jonaso@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#46009}