In CongestionControlFeedbackTracker save some history.
This allows to produce more accurate feedbacks in case of packet order
across produced feedback, as well as when information about already
received packet has changed. In particular it ensures that if packet was
reported as received, it would continue to be reported as received in a
new feedback.

The new code structure is also a preparation for implementation of the
RFC8888 based stats regarding number of packets reported as lost and
reported as recovered.

Bug: webrtc:374550342
Change-Id: I3d8ea116da3f7c0016479e7ad9ab8c3129b56766
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/418780
Reviewed-by: Per Kjellander <perkj@webrtc.org>
Commit-Queue: Danil Chapovalov <danilchap@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#46011}