Add support for EncoderSpeedExperiment in VideoStreamEncoder.
This CL populates the complexity setting in VideoCodec based on the
speed values set in WebRTC-EncoderSpeed field trial. An exception is
kept for low-core VP9 encoders, but this should be fazed out when the
dynamic speed controller is in place (it's still not used as of this
CL).

Bug: webrtc:443906251
Change-Id: I23a069f2de52bde1c30d5b44d5c047b668610877
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/415020
Commit-Queue: Erik Språng <sprang@webrtc.org>
Reviewed-by: Sergey Silkin <ssilkin@webrtc.org>
Reviewed-by: Ilya Nikolaevskiy <ilnik@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#46029}