Update VideoReceiveStream2::CalculateCorruptionScore to user RaceChecker
We should not DCHECK that we are on the decoder sequence, since that
might not be true when hardware decoders are used. However, the current
implementation is fine a long as there are no concurrent decoder
callbacks - and that should still hold true also for HW decoders.

Thus, switch to a `RaceChecker` instead of a `SequenceChecker`.

Bug: chromium:446727802
Change-Id: I22ed517e25936647b617ac8d49cc0b374f36d323
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/419362
Reviewed-by: Rasmus Brandt <brandtr@webrtc.org>
Reviewed-by: Johannes Kron <kron@webrtc.org>
Commit-Queue: Erik Språng <sprang@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#46028}