Implement `DecodabilitySimulator`.
The `DecodabilitySimulator` connects the previously landed components,
with the goal to provide a simulation of _decodable_ frames from an
RtcEventLog. The output sequence can be used for downstream data
analysis.

Implementation details:
* Simulation is driven by the `RtcEventLogDriver`.
* The stream specifics are handled by the new `DecodabilitySimulatorStream`. This one delegates to `Assembler` for frame assembly, `DecodabilityTracker` for decodability checks (through the `FrameBuffer`), and the new `DecodableFrameCollector` for keeping track of frame metadata.

The input is an RtcEventLog, and the output is a set of streams
containing simulated frames.

The upcoming `RenderingSimulator` will be very similar to the
`DecodabilitySimulator`, except that it will use the `RenderingTracker`
for rendering checks. It will therefore depend on the production jitter
buffer (`VideoStreamBufferController` and `IncomingVideoStream`).

Bug: b/423646186
Change-Id: I906acde2d2412d0518d8560b30bfbde91c52663c
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/419060
Reviewed-by: Åsa Persson <asapersson@webrtc.org>
Commit-Queue: Rasmus Brandt <brandtr@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#46051}