Implement `RenderingSimulator`.
This one is very similar to `DecodabilitySimulator`, except that it
relies on the production jitter buffer. That is, it provides a
simulation of _rendered_ frames from an RtcEventLog. The output sequence
can be used by the `event_log_visualizer` and other tools.

Implementation follows that of the `DecodabilitySimulator`.

Bug: b/423646186
Change-Id: Ib258eb8967d28f310a9e804ff775d041905e31bc
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/420100
Reviewed-by: Åsa Persson <asapersson@webrtc.org>
Commit-Queue: Rasmus Brandt <brandtr@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#46056}