Add EncoderSpeedController
This class dynamically adjusts encoder speed settings based on encode time and QP. It aims to find the optimal balance between CPU usage and video quality. The controller takes a configuration with a list of speed levels, each specifying speeds for key, base, middle, and top layers. It uses an exponential filter to track encode times and QP, increasing speed if overuse is detected or QP is very low, and decreasing speed if there is CPU headroom and QP is not too low.

This is the first part of a larger change to integrate dynamic speed control into the encoders.

Bug: webrtc:443906251
Change-Id: I0dcbe5d2ae16104a6ec00ed7c76629a605c22c17
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/419962
Reviewed-by: Ilya Nikolaevskiy <ilnik@webrtc.org>
Commit-Queue: Erik Språng <sprang@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#46084}