Support AbslStringify with VideoCodecType.
A downstream project wants to ASSERT that a webrtc codec type is what we
expect it to be. AbslStringify support is needed for failures to be easy
to read (otherwise it says a number is not equal to a number).

Bug: b/452368490
Change-Id: Ia261dd14645e7652f693aa8def0913137a47f1e3
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/422780
Auto-Submit: Henrik Boström <hbos@webrtc.org>
Reviewed-by: Henrik Andreassson <henrika@webrtc.org>
Commit-Queue: Henrik Andreassson <henrika@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#46106}