Update PipeWire tests to use Wireplumber instead PipeWire Media Session
Fix PipeWireStreamTest.TestPipeWire test as updating the framerate makes
the stream to be paused. We need to push a frame in order to restart it.

Also use newly added PipeWire 1.4 and do not run the tests on Linux-asan
as there are leaks in PipeWire caused by loading modules.

Fix for PipeWireStreamTest is submitted on behalf of Eliot Courtney.

Bug: chromium:414491370
Change-Id: Ia7b569604d452d0455a9f330fde0cecc6a2bc3fb
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/389040
Reviewed-by: Jeremy Leconte <jleconte@google.com>
Commit-Queue: Jan Grulich <grulja@gmail.com>
Reviewed-by: Alexander Cooper <alcooper@chromium.org>
Cr-Commit-Position: refs/heads/main@{#46128}