Downgrade unknown protocol check to a warning
When reporting the best connection state, the code increments
histograms based on the protocol of the local ICE candidate. If it
encountered a protocol for which no histogram was defined, it
would hit an `RTC_CHECK_NOTREACHED()`, causing a fatal crash.

This patch changes the fatal check to a non-fatal `LS_WARNING` log
message. This makes the statistics reporting more robust by
preventing a crash when an unexpected or unsupported protocol is
encountered, simply logging the event instead.

Bug: none
Change-Id: Ie6c25e126caaa972ffdaefebf30e93b55a535c68
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/422963
Reviewed-by: Harald Alvestrand <hta@webrtc.org>
Commit-Queue: Tomas Gunnarsson <tommi@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#46133}