Allow corruption score test where only one participant wants it.
Since webrtc:383078466 has been solved we can now allow this test. Even if one of the participants wants to activate corruption score this should not happen when the other participant does not want this. Previously this was failing, nevertheless, with the latest change this has been solved.

Bug: webrtc:383078466
Change-Id: Ia6beb6c9e3e8ff45168f84a7f6c970519f2097da
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/423080
Reviewed-by: Per Kjellander <perkj@webrtc.org>
Reviewed-by: Erik Språng <sprang@webrtc.org>
Commit-Queue: Emil Vardar (xWF) <vardar@google.com>
Cr-Commit-Position: refs/heads/main@{#46154}