Prepare to make the Network class immovable
This simplifies the reasoning about signals that are connected to the
Network class. The user of Network in a vector is changed to use a
vector of std::unique_ptr instead.

This CL leaves the copy operator in, so that downstream can be converted
to Clone() after landing this CL.

Bug: webrtc:42222066
Change-Id: Ifd1b4ea05ccd6a1a36af10a8a60bfaf02061408b
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/423281
Reviewed-by: Tomas Gunnarsson <tommi@webrtc.org>
Commit-Queue: Harald Alvestrand <hta@webrtc.org>
Reviewed-by: Danil Chapovalov <danilchap@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#46156}