Reinterpret back to the allocated type before delete
This modification addresses a bug detected in FreeWithSize (see
https://g-issues.chromium.org/issues/410190984). The existing
implementation involves Undefined Behavior, which results in the
compiler passing an incorrect size to the free function. 1. A C struct
object of 28 bytes is allocated using the new operator. 2. The object is
reinterpreted as a different C struct, which is 16 bytes. 3. The object
is then deleted, and the size is considered as 16 bytes.

Bug: chromium:410190984
Change-Id: I802892b26d133d6e98db6ffdc7ca0dfc05baddd4
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/425280
Commit-Queue: Ayumi Ono <ayumiohno@google.com>
Reviewed-by: Danil Chapovalov <danilchap@webrtc.org>
Reviewed-by: Harald Alvestrand <hta@webrtc.org>
Auto-Submit: Ayumi Ono <ayumiohno@google.com>
Commit-Queue: Harald Alvestrand <hta@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#46180}