Add a test for a synchronous (not asynchronous) TQ coroutine
This patch introduces a new test case, `CoroutineSynchronous`, to verify
the behavior of a coroutine that is intended to execute synchronously.

The coroutine, `ActuallySynchronous`, uses `co_await` to switch to the
already-current task queue. This specific usage should result in
immediate execution without introducing any asynchronous operations or
task posting, as the destination is already the current queue.

A related change is also included in the existing `Coroutine` test to
add an expectation that the coroutine's final callback is not invoked
immediately after the coroutine is launched, confirming its asynchronous
nature.

Bug: none
Change-Id: Ic9e122759029c8db077a6cafa925818e98e7da56
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/424982
Commit-Queue: Tomas Gunnarsson <tommi@webrtc.org>
Reviewed-by: Danil Chapovalov <danilchap@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#46178}