Refine thread blocking call guard implementation
The `ScopedDisallowBlockingCalls` class is updated to be null-safe and
only perform thread-specific operations when `Thread::Current()` is
available.

This change:
* Makes `ScopedDisallowBlockingCalls` robust when used on threads
    that do not have an associated `webrtc::Thread` object, preventing
    a potential crash (dereferencing a null `thread_`).
* Introduces the `RTC_DCHECK_DISALLOW_THREAD_BLOCKING_CALLS()` macro
    for convenience and to not affect non-dcheck enabled builds.
* Converts `std::function` to `absl::AnyInvocable` in the
    `ScopedCountBlockingCalls` constructor and destructor.
* Adds unit tests to cover the null-safe behavior and ensure the
    blocking call disallowance mechanism works as expected.
* Integrates the new disallow macro into `pc/codec_vendor.cc`.

The reason for introducing use of these checks to CodecVendor is because
it's where a lot of blocking calls are being issued from, in particular
via `PayloadTypeSuggester`. Addressing that is however not a part of
this change.

Bug: none
Change-Id: I03c1e4f93675b8297473953bfb1df0102666b388
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/425940
Reviewed-by: Harald Alvestrand <hta@webrtc.org>
Commit-Queue: Harald Alvestrand <hta@webrtc.org>
Auto-Submit: Tomas Gunnarsson <tommi@webrtc.org>
Commit-Queue: Tomas Gunnarsson <tommi@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#46202}