Add awareness of the signaling thread to JsepTransportController
Update JsepTransportController to accept the signaling thread as a
constructor dependency. This explicit dependency allows methods like
SetLocalDescription to assert execution on the signaling thread before
dispatching work to the network thread via BlockingCall.

Fix some tests to align with this and how the code runs in production.

Remove the internal initial_offerer_ state. The initial ICE role can be
initialized in SetLocalDescription without the initial_offerer_
variable. SdpOfferAnswerHandler has a state with that name and this
duplication in the controller shouldn't be necessary.

Bug: webrtc:360058654
Change-Id: I422a101677cf8332947919fb68e7868b3ca79708
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/426661
Reviewed-by: Harald Alvestrand <hta@webrtc.org>
Commit-Queue: Tomas Gunnarsson <tommi@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#46237}