Allow DatagramConnection (ie the RtcTransport web API) to send and receive real SRTP/SCTP
When wire protocol is set to kDtlsSrtp, take RTP/RTCP packets passed
to sendPackets and SRTP encrypt them in the same way as they would be
in a full PeerConnection, without the goofy wrapping in faked RTP
headers. Similarly decrypt received SRTP/SRTCP packets, decrypt them and
pass up to the client with the original headers inc seq num, ssrc etc.
For any other packets, just DTLS encrypt/decrypt, as in SrtpTransport.

This allows users of DatagramConnection to be wire-compatible with
a standard peerconnection.

Bug: chromium:443019066
Change-Id: I3bc3597b84b14fb03f3e7b6668638df0a5d86ef3
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/421782
Commit-Queue: Tony Herre <herre@google.com>
Reviewed-by: Harald Alvestrand <hta@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#46240}