Construct JsepTransportController on the signaling thread
Moves the instantiation of `JsepTransportController` from the network
thread to the signaling thread within `PeerConnection::Initialize`.
This involves preparing the configuration and creating the unique
pointer on the signaling thread, then moving the controller into the
network thread's blocking call for initialization.

Changers include updating `InitializeRtcpCallback` and
`InitializeUnDemuxablePacketHandler` to use `absl::AnyInvocable`
instead of `std::function`.

Deconstruction of JsepTransportController still occurs on the network
thread as that's still where ownership of the object is.

Bug: webrtc:360058654
Change-Id: I36163db5248e91df50d60030e0c017cf9208a00d
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/427140
Reviewed-by: Harald Alvestrand <hta@webrtc.org>
Commit-Queue: Tomas Gunnarsson <tommi@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#46252}