Replace BasicRegatheringController with RepeatingTaskHandle
The `BasicRegatheringController` class was essentially a thin wrapper
around a recurring task. This change removes the separate class and
implements the regathering logic directly in `P2PTransportChannel` using
`webrtc::RepeatingTaskHandle` which also avoids storing a pointer to the
PortAllocatorSession outside of P2PTransportChannel.

This simplifies the code by removing an extra layer of abstraction and
reducing boilerplate.

Key changes:
* Deleted `p2p/base/regathering_controller.{h,cc}` and the unit tests
  file.
* Migrated the tests `p2p/base/p2p_transport_channel_unittest.cc` to
  ensure continued coverage of regathering scheduling and intervals.
* Updated `P2PTransportChannel` to manage the recurring regathering
  task directly using `RepeatingTaskHandle`.

This will help allow initial construction of P2PTransportChannel on the
signaling thread in the future even though ownership and logic will
continue to run on the network thread. This also retains state within
the class, avoiding having more classes pointing to the same state.

Bug: none
Change-Id: I019f573cb3e6a52f1b46e01610a748f1691aacab
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/427681
Reviewed-by: Harald Alvestrand <hta@webrtc.org>
Commit-Queue: Tomas Gunnarsson <tommi@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#46282}