Cleanup field trial WebRTC-StunInterPacketDelay
Appears unused as no one is reading the weak_transport_interval_ var in
P2PTransportChannel. This looks like a bug, so we should either clean it
up, or pass it into the IceConfig ctor where kWeakPingInterval is used
as the default.

For now proposing to clean it up unless there is downstream usage.

Bug: webrtc:42221607
Change-Id: Ic2c559e68f9fc799f75da6599856528c9dface8e
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/427500
Commit-Queue: Joachim Reiersen <joachimr@meta.com>
Reviewed-by: Danil Chapovalov <danilchap@webrtc.org>
Reviewed-by: Per Kjellander <perkj@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#46313}