Remove redundant DtlsTransportInternal pointers from DtlsSrtpTransport
Remove the rtp_dtls_transport_ and rtcp_dtls_transport_ member variables
from DtlsSrtpTransport, which were redundant with the packet transport
pointers stored in the base RtpTransport class. DtlsSrtpTransport now
accesses the transports via the base class and casts them to
DtlsTransportInternal*, which is safe because SetDtlsTransports enforces
the type.

This change consolidates ownership and storage of the transport
pointers, reducing the risk of desynchronization and simplifying the
state management.

Bug: none
Change-Id: If6ad9340cea1532ac19d7672340d2abee56413af
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/429143
Reviewed-by: Harald Alvestrand <hta@webrtc.org>
Commit-Queue: Tomas Gunnarsson <tommi@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#46342}