Allow GoogCC to change default pacing factor if SendSide BWE is used
This change GoogCC to set a lower default pacing factor once it is confirmed that send side BWE is used regardless of if RFC 8888 or TWCC is used.
Per default now, pacing factor of 2.5 is used before first feedback, and 1.1 after first feedback.

If TWCC is used, the pacing factor is still overridden by RtpTransportController::SetPacingFactor. This method is unused unless TWCC is negotiated.





Bug: webrtc:464217874
Change-Id: I747261008df0ef0336cc70e11a187f03ebe810da
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/429402
Reviewed-by: Erik Språng <sprang@webrtc.org>
Commit-Queue: Per Kjellander <perkj@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#46364}