Duplicate the payload type picker on the signaling thread
This change introduces a new payload type suggester that runs solely
on the signaling thread.
It runs in parallel with the old PTP to check for errors; the plan
is to let it run in canary for a day, and then delete the old
payload type suggester in JsepTransportController.

The checking code and the old payload type suggester will be deleted in https://webrtc-review.googlesource.com/c/src/+/429344.

Bug: webrtc:360058654
Change-Id: Ia8ae7d173443b00efe9c7d057af5b07e252e4f11
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/429400
Commit-Queue: Tomas Gunnarsson <tommi@webrtc.org>
Auto-Submit: Harald Alvestrand <hta@webrtc.org>
Reviewed-by: Tomas Gunnarsson <tommi@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#46373}