Move `sim_time_impl` classes to separate target.
This change is the first in a small chain, where the goal is to
introduce a new class `SimulatedTimeTaskQueueController`, which will
provide `SimulatedTaskQueue`s that can be used from non-testonly
contexts. In particular, this will be used by the `RtcEventLogDriver` in the new video timing simulator. See full context in the Gerrit
relation chain.

The difference to `GlobalSimulatedTimeController` is two-fold:
1) The new class has no dependence on the process global clock.
2) [Due to 1)] The new class does not support creating
   `SimulatedThread`s.

This change moves some implementation details to a new target (no
functional changes intended). The upcoming change will introduce the new
class. And the third change in the chain will migrate the
`RtcEventLogDriver` onto the new class, setting `testonly=false` on the
corresponding target.

Bug: b/423646186
Change-Id: Ife21ec5d33a41d1b6e9e1536b88b36af2ac605bb
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/430940
Commit-Queue: Rasmus Brandt <brandtr@webrtc.org>
Reviewed-by: Jeremy Leconte <jleconte@google.com>
Cr-Commit-Position: refs/heads/main@{#46400}