Remove dependency on `channel->mid()` in stats collectors.
The mid is available on the transceivers instead. Using channel
properties in the stats code, complicates things a bit due to the
transceiver->channel hierarchy and the order of object creation. The mid
property is redundant on the channel object and should also be accessed
consistently on the network thread.

Also updates FakePeerConnectionForStats to ensure transceivers have MIDs
set when created.

Bug: none
Change-Id: I669a049e67a14106f0f0c4f17fa4d0ad86da345e
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/431380
Commit-Queue: Tomas Gunnarsson <tommi@webrtc.org>
Reviewed-by: Henrik Boström <hbos@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#46422}